Flesh out duck controller and client (#3)
The controller automatically creates appropriate CRDs for DuckType
resources along with ClusterRoles for view and edit permissions for the
duck.

A controller is started for each DuckType's CRD to check that the
referenced API exists and contribute RBAC rules to the view and edit
ClusterRoles.

The client defines a broker, which acts like an informer that is
watching all APIs registered for the duck type. Individual reconcilers
can subscribe to a broker and combine it with a tracker to be enqueue
requests when a tracked duck is modified.

The duck client should be used in favor to the regular client when an
object is known to be a duck. The client normalizes the version of the
object fetched to ensure the most compatible version of the resource for
the duck is used, and also reduces duplicate informers caused by version
skew. For example, informing on both v1 and v1beta1 for the same
resource. The shape of the functions mimick the client.Client interface,
but require additional metadata about the duck being fetched.

Signed-off-by: Scott Andrews <scott@andrews.me>