Duck client implements Client interface (#7)
The duck client now implemnets the controller-runtime Client interface
with the addition of tracking methods. New clients are created for a
specific duck type and then all methods operate against that type.

List and DeleteAllOf methods operate across all types implementing the
duck type unless a specific type is specified in the TypeMeta.
Internally, the client request is made for each implementing API
aggregating the result.

Signed-off-by: Scott Andrews <scott@andrews.me>