Expose SubManager as a reconciler
SubManagerReconciler starts a new sub manager based on the state of the
reconciled resource. Kinds to track in the submanager cache must be
explicitly defined in the LocalTypes func. All other client requests
are passed to the parent reconciler's client/cache.

The work performed by the nested manager is registered in the
SetupWithSubManager func.

The submanager is canceled when the resource that kicked off the sub
manager is deleted. Mutations to that resource are ignored once the sub
manager is started.

The resource must have a finalizer defined in order to the submanager to
cleanup after itself. The value of AssertFinalizer must exist on the
resource for the sub manager to start.

Signed-off-by: Scott Andrews <scott@andrews.me>